Reestructuracion Nocturna

+++ PrototipoBomberman/src/Juego/EstadoDestruible.java
@@ -7,42 +7,56 @@ public class EstadoDestruible extends EstadoCelda {
 
 	@Override
 	public void destruir(Celda c) {
+		c.matarBomberman();
 		c.destruirPersonajes();
 		c.setEstado(new EstadoTransitable());
-	}	
+	}
 
 	@Override
 	public void avanzar(Bomberman bomberman, Celda celdaSiguiente, int dir) {
 
-		if (bomberman.enModoDios()){
+		if (bomberman.enModoDios()) {
 			if (celdaSiguiente.hayEnemigos()) {
 				bomberman.matar();
-				System.out.println("El bomberman colisiona con un enemigo");			
-			}else{
+				System.out.println("El bomberman colisiona con un enemigo");
+			} else {
 				if (bomberman.getLock() == false) {
+					bomberman.lock();
 					bomberman.getCelda().eliminarBomberman();
 					bomberman.setCelda(celdaSiguiente);
-					celdaSiguiente.agregarBomberman(bomberman);					
+					celdaSiguiente.agregarBomberman(bomberman);
+					
 					bomberman.moverGrafica(dir);
-				}					
+				} else
+					System.out.println("Bomberman esta bloqueado");
 			}
 		}
 	}
 
 	@Override
-	public void avanzar(Enemigo enemigo, Celda celdaSiguiente) {
+	public void avanzar(Enemigo enemigo, Celda celdaSiguiente, int dir) {
 
-		if (enemigo.enModoDios()){
-			if (celdaSiguiente.hayBomberman()) {			
-				System.out.println("El enemigo colisiona con bomberman");
-				celdaSiguiente.matarBomberman();	
-				enemigo.getCelda().eliminarEnemigo(enemigo);
-				enemigo.setCelda(celdaSiguiente);
-				celdaSiguiente.agregarEnemigo(enemigo); //TODO Notitificar que murio bomberman y la logica sigue
+		if (enemigo.enModoDios()) {
+			if (celdaSiguiente.hayBomberman()) {
+				if (enemigo.getLock() == false) {
+					enemigo.lock();
+					System.out.println("El enemigo colisiona con bomberman");
+					celdaSiguiente.matarBomberman();
+					enemigo.getCelda().eliminarEnemigo(enemigo);
+					enemigo.setCelda(celdaSiguiente);
+					celdaSiguiente.agregarEnemigo(enemigo);
+					
+					enemigo.moverGrafica(dir);
+				}
 			} else {
-				enemigo.getCelda().eliminarEnemigo(enemigo);
-				enemigo.setCelda(celdaSiguiente);
-				celdaSiguiente.agregarEnemigo(enemigo);
+				if (enemigo.getLock() == false) {
+					enemigo.lock();
+					enemigo.getCelda().eliminarEnemigo(enemigo);
+					enemigo.setCelda(celdaSiguiente);
+					celdaSiguiente.agregarEnemigo(enemigo);
+					
+					enemigo.moverGrafica(dir);
+				}
 			}
 		}
 	}
